third_party/WebKit/LayoutTests/platform/win/external/wpt/web-animations/interfaces/Animation/playState-expected.txt

This is a testharness.js-based test.
PASS Animation.playState reports 'pending'->'running' when initially played 
PASS Animation.playState reports 'pending'->'paused' when pausing 
PASS Animation.playState is 'idle' when canceled. 
PASS Animation.playState is 'paused' after cancelling an animation, seeking it makes it paused 
Harness: the test ran to completion.
